requirements.txt
alembic==1.16.3
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
fastapi==0.116.0
fonttools==4.58.5
greenlet==3.2.3
h11==0.16.0
idna==3.10
kiwisolver==1.4.8
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
psycopg2-binaryboto3
